refactor: TaskTransaction works, will refactor, make input tasks a stream and then post.
