+++ src/python/dxl/cluster/database/tasks.py
@@ -1,11 +1,9 @@
 from flask import request
 from flask_restful import Api, Resource, reqparse
-from ..database import Task #, TaskTransactions
-from ..database.model import SlurmTask, Mastertask, taskSchema, slurmTaskSchema, masterTaskschema
+from ..database.model import Mastertask
 from ..backend import Backends
-# from ..interactive import post
 from ..database.transactions import deserialization, serialization
-# from ..interactive.web import post
+from ..interactive.transaction import TaskTransactions
 
 
 API_VERSION = 1
@@ -19,6 +17,7 @@ class TasksBind:
     Bind to transactions to given TaskTransactions, thus fixed database.
     """
     tasks = None
+
     @classmethod
     def set(cls, tasks: TaskTransactions):
         if cls.tasks is not None:
@@ -61,45 +60,29 @@ class TaskPoster:
         return {k: v for k, v in self.body["details"].items() if k not in ["is_user_task"]}
 
     @property
-    def is_user_task(self):
+    def is_master_task(self):
         try:
             return self.body["details"]["is_user_task"]
         except KeyError:
             print("Error! is_user_task field is requested!")
 
-    # def _on_independent(self, depends):
-    #     for d in depends:
-    #         pass
-    #         # Request.task
 
     def post(self):
         if self.backend is Backends.slurm:
-            # print(self.body)
-            # print(self.task_details)
-            # print(self.task_body)
-            # print(self.is_user_task)
-            # return {'finish': None,
-            #          'state': 'Created',
-            #          'submit': None,
-            #          'depends': [],
-            #          'id': 10,
-            #          'create': None,
-            #          'thenext': None}
-
             task = deserialization(self.task_body)
-            # print(type(task))
-            task = post(task)
+            task = TasksBind.tasks.post(task)
 
             slurmTask = deserialization(self.task_details)
             slurmTask.task_id = task.id
-            slurmTask = post(slurmTask)
+            slurmTask = TasksBind.tasks.post(slurmTask, task.id)
 
-            if self.is_user_task:
-                masterTask = post(Mastertask(backend=Backends.slurm.value, backend_task_id=slurmTask.id))
+            if self.is_master_task:
+                masterTask = TasksBind.tasks.post(Mastertask(backend=Backends.slurm.value, id=slurmTask.id))
                 return serialization(masterTask), 200
 
             return serialization(slurmTask), 200
 
 
-def add_resource(api):
+def add_resource(api, tasks):
+    TasksBind.set(tasks)
     api.add_resource(TasksResource, TASK_API_URL)  # /api/v1/tasks
